+++ webkit/glue/plugins/npapi_extension_thunk.cc
@@ -62,7 +62,7 @@ static NPError Device2DInitializeContext(NPP id,
 static NPError Device2DSetStateContext(NPP id,
                                        NPDeviceContext* context,
                                        int32 state,
-                                       int32 value) {
+                                       intptr_t value) {
   scoped_refptr<NPAPI::PluginInstance> plugin = FindInstance(id);
   if (plugin) {
     return plugin->webplugin()->delegate()->Device2DSetStateContext(
@@ -74,7 +74,7 @@ static NPError Device2DSetStateContext(NPP id,
 static NPError Device2DGetStateContext(NPP id,
                                        NPDeviceContext* context,
                                        int32 state,
-                                       int32* value) {
+                                       intptr_t* value) {
   scoped_refptr<NPAPI::PluginInstance> plugin = FindInstance(id);
   if (plugin) {
     return plugin->webplugin()->delegate()->Device2DGetStateContext(
@@ -174,7 +174,7 @@ static NPError Device3DInitializeContext(NPP id,
 static NPError Device3DSetStateContext(NPP id,
                                        NPDeviceContext* context,
                                        int32 state,
-                                       int32 value) {
+                                       intptr_t value) {
   scoped_refptr<NPAPI::PluginInstance> plugin = FindInstance(id);
   if (plugin) {
     return plugin->webplugin()->delegate()->Device3DSetStateContext(
@@ -186,7 +186,7 @@ static NPError Device3DSetStateContext(NPP id,
 static NPError Device3DGetStateContext(NPP id,
                                        NPDeviceContext* context,
                                        int32 state,
-                                       int32* value) {
+                                       intptr_t* value) {
   scoped_refptr<NPAPI::PluginInstance> plugin = FindInstance(id);
   if (plugin) {
     return plugin->webplugin()->delegate()->Device3DGetStateContext(
@@ -293,7 +293,7 @@ static NPError DeviceAudioInitializeContext(NPP id,
 static NPError DeviceAudioSetStateContext(NPP id,
                                           NPDeviceContext* context,
                                           int32 state,
-                                          int32 value) {
+                                          intptr_t value) {
   scoped_refptr<NPAPI::PluginInstance> plugin = FindInstance(id);
   if (plugin) {
     return plugin->webplugin()->delegate()->DeviceAudioSetStateContext(
@@ -305,7 +305,7 @@ static NPError DeviceAudioSetStateContext(NPP id,
 static NPError DeviceAudioGetStateContext(NPP id,
                                           NPDeviceContext* context,
                                           int32 state,
-                                          int32* value) {
+                                          intptr_t* value) {
   scoped_refptr<NPAPI::PluginInstance> plugin = FindInstance(id);
   return plugin->webplugin()->delegate()->DeviceAudioGetStateContext(
       static_cast<NPDeviceContextAudio*>(context), state, value);
